Fix compilation bug with local GSL installation.
With a local GSL installation (i.e. when overwritting the GSL_CONFIG
variable in ccmake), the GSL include files were not found at compilation
time.  This was because Cmake did not find properly the GSL include
directory at configuration time. This was due to the GSL_INCLUDE_DIR
variable being set to cache in FindGSL.cmake.